add permissions

+++ controllers/heroes.js
@@ -2,7 +2,6 @@ const { exit } = require('process');
 const mongoDB = require('../dbconnect');
 const ObjectId = require('mongodb').ObjectId;
 
-
 //heartbeat
 function naviagationUi(req, res) {
 
@@ -91,42 +90,45 @@ async function getHero(req, res) {
 async function createNewHero(req, res) {
 
     try {
+        if (await getPrivData(req.oidc.user.sub,'create')) {
+            //get new hero data from request object
+            const newHero = req.body;
+            // check object correctness
+            if (allKeysExist(heroTemplate, newHero) === false) {
+                res.status(400).send("Bad data.");
+                return;
+            }
 
-         //get new hero data from request object
-        const newHero = req.body;
-        // check object correctness
-        if (allKeysExist(heroTemplate, newHero) === false) {
-             res.status(400).send("Bad data.");
-            return;
+            //get db
+            const db = mongoDB.getDB().db("heroes");
+
+            //get new id number
+            const newId = await db.collection("heroes").find(
+                {},
+                { projection: { id: 1, _id: 0 } })
+                .sort({ id: -1 })
+                .limit(1).toArray();
+
+            newHero['id'] = newId[0]['id'] + 1;
+            
+
+            // create user in database
+            await db.collection("heroes").insertOne(newHero, (err, result) => {
+                if (err) {
+                    console.log(err);
+                    return err;
+                }
+                else {
+                    setHeaders(res);
+                    delete result['insertedId'];
+                    //return new id number
+                    result['id'] = newHero['id'];
+                    res.status(201).send(result);
+                }
+            });
+        } else {
+            res.status(403).send("Not Authorized");    
         }
-
-        //get db
-        const db = mongoDB.getDB().db("heroes");
-
-        //get new id number
-        const newId = await db.collection("heroes").find(
-            {}, 
-            { projection: { id: 1, _id: 0} })
-            .sort({ id: -1})
-            .limit(1).toArray();
-
-        newHero['id'] = newId[0]['id'] + 1;
-        
-
-        // create user in database
-        await db.collection("heroes").insertOne(newHero, (err, result) => {
-            if (err) {
-                console.log(err);
-                return err;
-            }
-            else  {
-                setHeaders(res);
-                delete result['insertedId'];
-                //return new id number
-                result['id'] = newHero['id'];
-                res.status(201).send(result);
-            }
-        });
     } catch (err) {
         console.log(err);
         res.status(500).send(err);
@@ -137,38 +139,42 @@ async function createNewHero(req, res) {
 async function updateHero(req, res) {
 
     try {
-        //get id from request object
-        const heroId = parseInt(req.params.id);
+        if (await getPrivData(req.oidc.user.sub, 'update')) {
+            //get id from request object
+            const heroId = parseInt(req.params.id);
+            
+            //get new hero data from request object
+            const updatedHero = req.body;
+
+            // check object correctness
+            if (allKeysExist(heroTemplate, updatedHero) === false) {
+                res.status(400).send("Bad data.");
+                return;
+            }
         
-        //get new hero data from request object
-        const updatedHero = req.body;
 
-        // check object correctness
-        if (allKeysExist(heroTemplate, updatedHero) === false) {
-            res.status(400).send("Bad data.");
-            return;
-        }
-       
-
-        // if the hero id is missing from the data, add it.
-        if (updatedHero['id'] == null) {
-            updatedHero['id'] = heroId;
-        }
-
-        //update contact
-        const db = mongoDB.getDB().db("heroes");
-        await db.collection("heroes").replaceOne({ id: heroId }, updatedHero, (err, result) => {
-
-            if (err) {
-                console.log(err);
-                return err;
+            // if the hero id is missing from the data, add it.
+            if (updatedHero['id'] == null) {
+                updatedHero['id'] = heroId;
             }
-            else  {
-                // return result
-                setHeaders(res);
-                res.status(204).send('');
-            }
-        })
+
+            //update contact
+            const db = mongoDB.getDB().db("heroes");
+            await db.collection("heroes").replaceOne({ id: heroId }, updatedHero, (err, result) => {
+
+                if (err) {
+                    console.log(err);
+                    return err;
+                }
+                else {
+                    // return result
+                    setHeaders(res);
+                    res.status(204).send('');
+                }
+            })
+        } else {
+            res.status(403).send("Not Authorized");    
+        }
      } catch (err) {
         console.log(err);
         res.status(500).send(err);
@@ -179,26 +185,29 @@ async function updateHero(req, res) {
 async function deleteHero(req, res) {
 
     try {
-
-        //get id from request object
-        const heroId = parseInt(req.params.id);
-        
-        //get data
-        const dbo = mongoDB.getDB().db("heroes");
-        await dbo.collection("heroes").deleteOne({ id: heroId }, (err, result) => {
-            if (err)
-                return err;
-            else
-                if (result.deletedCount == 0) {
-                    setHeaders(res);
-                    res.status(404).send(result);
-                } else {
-                    //return data
-                    setHeaders(res);
-                    res.status(200).send(result);
-                }
-        });
-    } catch (err) {
+        if (await getPrivData(req.oidc.user.sub, 'delete')) {
+            //get id from request object
+            const heroId = parseInt(req.params.id);
+            
+            //get data
+            const dbo = mongoDB.getDB().db("heroes");
+            await dbo.collection("heroes").deleteOne({ id: heroId }, (err, result) => {
+                if (err)
+                    return err;
+                else
+                    if (result.deletedCount == 0) {
+                        setHeaders(res);
+                        res.status(404).send(result);
+                    } else {
+                        //return data
+                        setHeaders(res);
+                        res.status(200).send(result);
+                    }
+            });
+        } else {
+            res.status(403).send("Not Authorized");    
+        }
+        } catch (err) {
         console.log(err);
         res.status(500).send(err);
     }
@@ -213,6 +222,32 @@ function setHeaders(res) {
     res.setHeader('Access-Control-Allow-Credentials', true);
     
 }
+
+async function canCreate(sid) {
+    const user = await getPrivData(sid);
+    if (user != null) {
+        return user.privileges.create;
+    } else {
+        console.log("user not found");
+        return false;
+    }
+}
+
+
+async function getPrivData(sub,priv) {
+
+    const dbo = await mongoDB.getDB().db("heroes");
+    const user = await dbo.collection("privileges").findOne({ user_id: sub });
+    if (user != null) {
+        return user.privileges[priv];
+    } else {
+        console.log("User not found.");
+        return false;
+    }
+
+    return returnValue;
+}
+
 
 
 function allKeysExist(template, newHero) {
